+++ steam-plus.meta.js
@@ -10,7 +10,7 @@
 // @name:ko            Steam Plus
 // @name:pl            Steam Plus
 // @namespace          https://github.com/NemoKing1210/steam-plus
-// @version            0.14.1
+// @version            0.16.2
 // @author             NemoKing1210
 // @description        Improves Steam functionality
 // @description:ru     Улучшает функционал Steam